+++ src/vs/workbench/services/workingCopy/common/workingCopyHistoryService.ts
@@ -4,23 +4,23 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { localize } from 'vs/nls';
-import { Emitter } from 'vs/base/common/event';
+import { Event, Emitter } from 'vs/base/common/event';
 import { assertIsDefined } from 'vs/base/common/types';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions } from 'vs/workbench/common/contributions';
-import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
+import { ILifecycleService, LifecyclePhase, WillShutdownEvent } from 'vs/workbench/services/lifecycle/common/lifecycle';
 import { WorkingCopyHistoryTracker } from 'vs/workbench/services/workingCopy/common/workingCopyHistoryTracker';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { IWorkingCopyHistoryEntry, IWorkingCopyHistoryEntryDescriptor, IWorkingCopyHistoryEvent, IWorkingCopyHistoryService, MAX_PARALLEL_HISTORY_IO_OPS } from 'vs/workbench/services/workingCopy/common/workingCopyHistory';
 import { FileOperationError, FileOperationResult, IFileService, IFileStatWithMetadata } from 'vs/platform/files/common/files';
 import { IRemoteAgentService } from 'vs/workbench/services/remote/common/remoteAgentService';
 import { URI } from 'vs/base/common/uri';
-import { DeferredPromise, Limiter } from 'vs/base/common/async';
+import { DeferredPromise, Limiter, RunOnceScheduler } from 'vs/base/common/async';
 import { dirname, extname, isEqual, joinPath } from 'vs/base/common/resources';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
 import { hash } from 'vs/base/common/hash';
 import { indexOfPath, randomPath } from 'vs/base/common/extpath';
-import { CancellationToken } from 'vs/base/common/cancellation';
+import { CancellationToken, CancellationTokenSource } from 'vs/base/common/cancellation';
 import { ResourceMap } from 'vs/base/common/map';
 import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
 import { ILabelService } from 'vs/platform/label/common/label';
@@ -784,6 +784,84 @@ export abstract class WorkingCopyHistoryService extends Disposable implements IW
 
 	protected abstract getModelOptions(): IWorkingCopyHistoryModelOptions;
 
+}
+
+export class NativeWorkingCopyHistoryService extends WorkingCopyHistoryService {
+
+	private static readonly STORE_ALL_INTERVAL = 5 * 60 * 1000; // 5min
+
+	private readonly isRemotelyStored = typeof this.environmentService.remoteAuthority === 'string';
+
+	private readonly storeAllCts = this._register(new CancellationTokenSource());
+	private readonly storeAllScheduler = this._register(new RunOnceScheduler(() => this.storeAll(this.storeAllCts.token), NativeWorkingCopyHistoryService.STORE_ALL_INTERVAL));
+
+	constructor(
+		@IFileService fileService: IFileService,
+		@IRemoteAgentService remoteAgentService: IRemoteAgentService,
+		@IWorkbenchEnvironmentService environmentService: IWorkbenchEnvironmentService,
+		@IUriIdentityService uriIdentityService: IUriIdentityService,
+		@ILabelService labelService: ILabelService,
+		@ILifecycleService private readonly lifecycleService: ILifecycleService,
+		@ILogService logService: ILogService,
+		@IConfigurationService configurationService: IConfigurationService
+	) {
+		super(fileService, remoteAgentService, environmentService, uriIdentityService, labelService, logService, configurationService);
+
+		this.registerListeners();
+	}
+
+	private registerListeners(): void {
+		if (!this.isRemotelyStored) {
+
+			// Local: persist all on shutdown
+			this.lifecycleService.onWillShutdown(e => this.onWillShutdown(e));
+
+			// Local: schedule persist on change
+			this._register(Event.any(this.onDidAddEntry, this.onDidChangeEntry, this.onDidReplaceEntry, this.onDidRemoveEntry)(() => this.onDidChangeModels()));
+		}
+	}
+
+	protected getModelOptions(): IWorkingCopyHistoryModelOptions {
+		return { flushOnChange: this.isRemotelyStored /* because the connection might drop anytime */ };
+	}
+
+	private onWillShutdown(e: WillShutdownEvent): void {
+
+		// Dispose the scheduler...
+		this.storeAllScheduler.dispose();
+		this.storeAllCts.dispose(true);
+
+		// ...because we now explicitly store all models
+		e.join(this.storeAll(e.token), { id: 'join.workingCopyHistory', label: localize('join.workingCopyHistory', "Saving local history") });
+	}
+
+	private onDidChangeModels(): void {
+		if (!this.storeAllScheduler.isScheduled()) {
+			this.storeAllScheduler.schedule();
+		}
+	}
+
+	private async storeAll(token: CancellationToken): Promise<void> {
+		const limiter = new Limiter(MAX_PARALLEL_HISTORY_IO_OPS);
+		const promises = [];
+
+		const models = Array.from(this.models.values());
+		for (const model of models) {
+			promises.push(limiter.queue(async () => {
+				if (token.isCancellationRequested) {
+					return;
+				}
+
+				try {
+					await model.store(token);
+				} catch (error) {
+					this.logService.trace(error);
+				}
+			}));
+		}
+
+		await Promise.all(promises);
+	}
 }
 
 // Register History Tracker
